% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_events.R
\name{add_missing_events}
\alias{add_missing_events}
\title{Identify events missing from the dataset and, if their expected event date has passed, add them to the data.}
\usage{
add_missing_events(
  dataset,
  event_data,
  condition = NULL,
  allocation = TRUE,
  site = TRUE
)
}
\arguments{
\item{dataset}{A REDCap export dataset.}

\item{event_data}{A CSV file containing the events and their associated anchors - template located in Github named "20240109_AddMissingEventsTemplate.csv".}

\item{condition}{Optional parameter that can be applied, e.g. if the events for intervention and control participants are different. Default is NULL.}

\item{allocation}{Logical, if allocations are used in the study then this should be true. Default to TRUE.}

\item{site}{Logical, if sites are used in the study then this should be true. Default to TRUE.}
}
\value{
A data frame containing missing events.
}
\description{
This function adds a row for any event that is expected in the data and is missing. NOTE this script should be run after allocation_data_cleaning and before encode_site_names.
}
\examples{
:
Example usage:
intervention_events <- read.csv("20240108_ExpectedEventsInvervention_V1.0.csv")
condition <- "Allocation == 'Intervention'"

complete_dataset <- add_missing_events(dataset, event_data, condition)

}
