% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_data_cleaning.R
\name{allocation_data_cleaning}
\alias{allocation_data_cleaning}
\title{Apply coded allocations to every row}
\usage{
allocation_data_cleaning(data, allocation_column, allocation_event_name)
}
\arguments{
\item{data}{A REDcap export dataset.}

\item{allocation_column}{Column name where allocation data is stored.}

\item{allocation_event_name}{Event name where allocations are assigned to participants.}
}
\value{
A dataframe containing allocations.
}
\description{
This function applies allocations to all rows for a participant
}
\examples{
:
Example useage:
dataset <- allocation_data_cleaning(data, "pdallocation", "randomisation_arm_1")


}
