% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_completeness.R
\name{calculate_completeness}
\alias{calculate_completeness}
\title{Calculates the completeness of a timepoint based on the \emph{timepoint_name}_compelte column values.}
\usage{
calculate_completeness(row)
}
\arguments{
\item{row}{A row defined from a dataset.}
}
\value{
A value indicated the completeness of the timepoint.
}
\description{
Outputs "Complete", "Partially complete", "Not started or "Undefined" as the completeness value for a row in a defined dataset.Used in conjundtion with plot_crf_completeness_data_preparation function.
}
\examples{

completeness_values <- apply(subset_table[, completeness_columns], 1, calculate_completeness)

}
