% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_k_anonymity.R
\name{calculate_k_anonymity}
\alias{calculate_k_anonymity}
\title{Calculates the k-anonymity score for a dataset based on quasi-identifiers}
\usage{
calculate_k_anonymity(dataset, quasi_identifiers)
}
\arguments{
\item{dataset}{Dataset to be assessed.}

\item{quasi_identifiers}{Vector defining the quasi-identifiers in the dataset}
}
\value{
A k-anonymity score
}
\description{
Calculates the k-anonymity score for a dataset based on quasi-identifiers
}
\examples{
:
Example usage:

quasi_identifiers <- c("age", "place_of_birth", "gender")
k_anonymity_value <- calculate_k_anonymity(dataset, quasi_identifiers)
}
