% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_l_diversity.R
\name{calculate_l_diversity}
\alias{calculate_l_diversity}
\title{Calculates the l-diversity score for a dataset based on quasi-identifiers and sensitive attributes.}
\usage{
calculate_l_diversity(dataset, quasi_identifiers, sensitive_attributes)
}
\arguments{
\item{dataset}{Dataset to be assessed.}

\item{quasi_identifiers}{Vector defining the quasi-identifiers in the dataset}

\item{sensitive_attributes}{Vector defining all of the sensitive attributes in a dataset.}
}
\value{
A a-diversity score
}
\description{
Calculates the l-diversity score for a dataset based on quasi-identifiers and sensitive attributes.
}
\examples{
:
Example usage:

quasi_identifiers <- c("age", "place_of_birth", "gender")
sensitive_attributes <- c("conviction_reason", "suicidal_thoughts") 
l_diversity_value <- calculate_l_diversity(dataset, quasi_identifiers, sensitive_attributes)

}
