% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorise_age_from_dob.R
\name{categorise_age_from_dob}
\alias{categorise_age_from_dob}
\title{Categorise age data in a dataset into age ranges (for anonymisation) from a date of birth column}
\usage{
categorise_age_from_dob(dataset, dob_column, breaks, labels)
}
\arguments{
\item{dataset}{Dataset containing the age field.}

\item{breaks}{Size of age category brackets.}

\item{labels}{Labels of age categories.}

\item{age_column}{Name of date of birth column}
}
\value{
an age column with categorised ages.
}
\description{
Categorise age data in a dataset into age ranges (for anonymisation) from a date of birth column
}
\examples{
:
Example usage:

breaks <- c(0, 20, 30, 40, 50, 60, 70, 80, 90, 99)
labels <- c("0-20", "20-30", "30-40", "40-50", "50-60", "60-70", "70-80", "80-90", "90-99")
anonymised_dataset <- categorise_ages(dataset, dmdob, breaks, labels)
}
