% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorise_ages.R
\name{categorise_ages}
\alias{categorise_ages}
\title{Categorise age data in a dataset into age ranges (for anonymisation)}
\usage{
categorise_ages(dataset, age_column, breaks, labels)
}
\arguments{
\item{dataset}{Dataset containing the age field.}

\item{age_column}{Name of age column to be categorised.}

\item{breaks}{Size of age category brackets.}

\item{labels}{Labels of age categories.}
}
\value{
an age column with categorised ages.
}
\description{
Categorise age data in a dataset into age ranges (for anonymisation)
}
\examples{
:
Example usage:

breaks <- c(0, 20, 30, 40, 50, 60, 70, 80, 90, 99)
labels <- c("0-20", "20-30", "30-40", "40-50", "50-60", "60-70", "70-80", "80-90", "90-99")
anonymised_dataset <- categorise_ages(dataset, dmage, breaks, labels)
}
