% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crf_completeness_data_preparation.R
\docType{data}
\name{category}
\alias{category}
\title{Generate a dataset of timepoint completeness to create a completeness plot.}
\format{
An object of class \code{character} of length 1.
}
\usage{
category
}
\arguments{
\item{data}{A complete dataset.}

\item{timepoint_names}{A vector containing the list of timepoint names.}

\item{category}{The category by which data will be organised, must be "Site" or "Allocation".}

\item{api_token}{The API token for accessing REDCap.}

\item{test}{Logical, indicating whether to use the test environment (default is FALSE).}
}
\value{
A dataset summarising event completeness.
}
\description{
This function produces a dataset which gives the completeness of each defined timepoint for each participant.
}
\examples{
timepoint_names <- c("week_0_blinded_ass_arm_1", "week_2_blinded_ass_arm_1", "week_6_blinded_ass_arm_1", "week_8_blinded_as_arm_1", "week_22_blinded_as_arm_1")

plot_crf_completeness_dataset_preparation <- function(dataset, timepoint_names, "Allocation", your_api_token, test=FALSE)

}
\keyword{datasets}
