% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_field_completeness.R
\name{check_field_completeness}
\alias{check_field_completeness}
\title{Check Field Completeness of a Data Frame}
\usage{
check_field_completeness(data, by = NULL, full = FALSE)
}
\arguments{
\item{data}{A data frame to check the completeness of its fields.}

\item{by}{The column name as a string to group the data by before checking field completeness. This argument is optional.}

\item{full}{A boolean that determines whether to return completeness information for missing values and a set of predefined codes. This argument is optional, and its default value is \code{FALSE}.}
}
\value{
A data frame with the completeness of each field.
}
\description{
This function calculates the completeness of each field in a given data frame.
}
