% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_kmean.R
\name{clustering_analysis}
\alias{clustering_analysis}
\title{Run k-means clustering for multivariate groups}
\usage{
clustering_analysis(
  data,
  metadata,
  multivariate_group_column = "multivariate_group",
  multivariate_yn_column = "multivariate_yn",
  variable_type_column = "variable_type",
  field_name_column = "field_name",
  record_id_column = "record_id",
  site_column = "redcap_data_access_group",
  event_column = "redcap_event_name",
  min_n = 10,
  k = 2,
  nstart = 25,
  site_prop_threshold = 0.8
)
}
\arguments{
\item{data}{A REDCap export dataset.}

\item{metadata}{A critical data item metadata dataframe.}

\item{multivariate_group_column}{Column name containing multivariate group names.}

\item{multivariate_yn_column}{Column name indicating whether a variable is multivariate.}

\item{variable_type_column}{Column name containing variable type.}

\item{field_name_column}{Column name containing REDCap field names.}

\item{record_id_column}{Column name containing participant/record ID.}

\item{site_column}{Column name containing site/group.}

\item{event_column}{Column name containing REDCap event name.}

\item{min_n}{Minimum number of complete records required to run clustering.}

\item{k}{Number of clusters to use.}

\item{nstart}{Number of random starts for k-means.}

\item{site_prop_threshold}{Proportion threshold used to flag site-dominated clusters.}
}
\value{
A list containing cluster-level outputs.
}
\description{
This function runs k-means clustering for continuous variables within each
multivariate group defined in the metadata. It is intended as an exploratory
analysis to identify unusual clustering patterns or site-dominated clusters.
}
\details{
Variables selected for clustering are coerced to numeric before complete-case
filtering. Records containing missing or non-finite values in any clustering
variable are excluded from the analysis.
}
\examples{
cluster_outputs <- clustering_analysis(data, metadata)

}
