% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_date_column.R
\name{create_date_column}
\alias{create_date_column}
\title{Takes partial date fields and creates a single date field}
\usage{
create_date_column(
  dataset,
  year_column,
  month_column,
  day_columns,
  new_date_column,
  is_month_numeric = FALSE
)
}
\arguments{
\item{dataset}{The dataset containing the date fields.}

\item{year_column}{The column where the year is specified.}

\item{month_column}{The column where the month is specified.}

\item{new_date_column}{The name of the new column containing the date.}

\item{is_month_numeric}{Logical, allows the user to specify if the month is given numerically (TRUE) or as words (e.g. June; FALSE). Default is FALSE.}

\item{day_colummns}{A vector where the day columns are specified.}
}
\value{
A column containing the combined date.
}
\description{
This function takes the fields used to make a partial date field and combines them to give a field in date format
}
\examples{
# Example usage:
day_columns <- c("pdday1", "pdday2", "pdday2", "pdday4")
dataset <- create_date_column(dataset, "pdyear", "pdmonth", day_columns, "pddat", is_month_numeric = FALSE)

}
