% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range_validation_dynamic.R
\name{date_range_validation_dynamic}
\alias{date_range_validation_dynamic}
\title{Author: Paigan Aspinall
Date & version: 10FEB2024 V1.0.1
R version: 4.2.2}
\usage{
date_range_validation_dynamic(dataset, rules)
}
\arguments{
\item{dataset}{A complete dataset.}

\item{rules}{A CSV file that is imported and contains all of the date range checks to be performed. Use template STU_DM_023_DateValidationTemplate_dynamic.R}
}
\value{
A dataset summarising potentially invalid date data.
}
\description{
Generate a dataset that returns all instances where the date validation rules defined in an external CSV are returned.
}
\details{
Produces a dataset of potentially invalid date data.
}
\examples{


date_validation_output_dynamic <- date_range_validation(dataset, rules)

}
