% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range_validation_repeating.R
\name{date_range_validation_repeating}
\alias{date_range_validation_repeating}
\title{Author: Paigan Aspinall
Date & version: 19FEB2024 V1.0.0
R version: 4.2.2}
\usage{
date_range_validation_repeating(dataset, rules)
}
\arguments{
\item{dataset}{A complete dataset.}

\item{rules}{A CSV file that is imported and contains all of the date range checks to be performed. Created using the template: DateValidationTemplateRepeating_V1.0}
}
\value{
A dataset summarising potentially invalid date data.
}
\description{
Generate a dataset that returns all instances where the date validation rules defined in an external CSV are returned.Note that both dates being checked must be within
the repeating instrument, otherwise the regular date validation functions can be used.
}
\details{
Produces a dataset of potentially invalid dates from dates within repeating instruments.
}
\examples{
date_validation_output <- date_range_validation_repeating(dataset, rules)

}
