% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range_validation_standard.R
\name{date_range_validation_standard}
\alias{date_range_validation_standard}
\title{Author: Paigan Aspinall
Date & version: 08FEB2024 V1.0.1
R version: 4.2.2}
\usage{
date_range_validation_standard(dataset, rules)
}
\arguments{
\item{dataset}{A complete dataset.}

\item{rules}{A CSV file that is imported and contains all of the date range checks to be performed. Created using the template: DateValidationTemplate_V1.0}
}
\value{
A dataset summarising potentially invalid date data.
}
\description{
Generate a dataset that returns all instances where the date validation rules defined in an external CSV are returned.
}
\details{
Produces a dataset of potentially invalid date data.
}
\examples{
date_validation_output_standard <- date_range_validation(dataset, rules)

}
