% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_since.R
\name{days_since}
\alias{days_since}
\title{Creates a new column in the dataset, which is the time difference (in days) between two given dates.}
\usage{
days_since(dataset, date_column, start_date)
}
\arguments{
\item{dataset}{Dataset to be assessed.}

\item{date_column}{Later date column, which will be deleted and replaced by a number of days}

\item{start_date}{Start date column, for example baseline, which will act as day 0.}
}
\value{
A dataset with a date column in days rather than in date format
}
\description{
Creates a new column in the dataset, which is the time difference (in days) between two given dates.
}
\examples{
:
Example usage:
dataset_clean <- days_since(dataset, week_1_dat, baseline_dat)
}
