% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_site_names.R
\name{encode_site_names}
\alias{encode_site_names}
\title{Add site names to the dataset}
\usage{
encode_site_names(dataset, api_token, test = FALSE)
}
\arguments{
\item{dataset}{The dataset where the new column is being added.}

\item{api_token}{The API token for accessing REDCap.}

\item{test}{Logical, indicating whether to use the test environment (default is FALSE).}
}
\value{
A data frame containing the new column.
}
\description{
This function retrieves DAG data from REDcap and adds a Site column to the dataset which contains the DAG name.
}
\details{
This function sends an HTTP POST request to the REDCap API to retrieve data.
It will then add items from this REDcap data to the specified dataset.
}
\examples{
# Example usage:
data <- get_redcap_data(your_dataset, api_token = "your_api_token", test = TRUE)

}
