% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_crf_data_from_RCC.R
\name{export_crf_data_from_RCC}
\alias{export_crf_data_from_RCC}
\title{Export CRF Data from RCC}
\usage{
export_crf_data_from_RCC(crf_name, token, env)
}
\arguments{
\item{crf_name}{The name of the CRF to export data from.}

\item{token}{The access token for authentication with RCC.}
}
\value{
A dataframe containing the exported CRF data.
}
\description{
This function exports data from the RedCap Cloud (RCC) system for a specific
CRF (Case Report Form) using the provided CRF name and access token.
}
\examples{
export_crf_data_from_RCC("CRF_Name", "access_token")

}
