% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_annotation_labels.R
\name{field_annotation_labels}
\alias{field_annotation_labels}
\title{Add variable labels to data from field annotations column}
\usage{
field_annotation_labels(dataset, data_dictionary)
}
\arguments{
\item{dataset}{The dataset to which the labels are being added}

\item{data_dictionary}{The dataset which contains the field_annotation column}
}
\value{
A dataframe with labelled columns
}
\description{
Add variable labels to data from field annotations column
}
\details{
This function formats the field annotation column to remove action tags then loops through the data dictionary to assign these field annotations as labels to the dataset.
}
\examples{
labelled_dataset <- field_annotation_labels(your_dataset, your_data_dictionary)

}
