% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_form_validations.R
\name{get_form_validations}
\alias{get_form_validations}
\title{Pulls all of the form validation checks that need to be performe dout of the data dictionary.}
\usage{
get_form_validations(api_token, test = FALSE)
}
\arguments{
\item{api_token}{The API token for accessing REDCap.}

\item{test}{Logical, indicating whether to use the test environment (default is FALSE).}
}
\value{
A csv containing all of the form validation checks.
}
\description{
This function retrieves data from REDCap data dictionary using the API and identifies all of the form validation checks required for the study.
}
\examples{
# Example usage:
get_form_validations(your_token, test=TRUE)

}
