% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redcap_audit.R
\name{get_redcap_audit}
\alias{get_redcap_audit}
\title{Import the REDCap audit log data using API token.}
\usage{
get_redcap_audit(api_token, test = FALSE)
}
\arguments{
\item{api_token}{The API token for accessing REDCap.}

\item{test}{Logical, indicating whether to use the test environment (default is FALSE).}
}
\value{
A data frame containing the requested data.
}
\description{
This function retrieves audit log data from REDCap using the REDCap API and an API token.
}
\examples{
:
Example usage:

audit_data <- get_redcap_audit("your_api_token", test=TRUE)

}
