% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redcap_data.R
\name{get_redcap_data}
\alias{get_redcap_data}
\title{Get data from REDCap using an API token}
\usage{
get_redcap_data(
  api_token,
  test = FALSE,
  labelVariables = TRUE,
  labelValues = FALSE
)
}
\arguments{
\item{api_token}{The API token for accessing REDCap.}

\item{test}{Logical, indicating whether to use the test environment (default is FALSE).}

\item{labelVariables}{Logical, indicating whether to label variables (default is TRUE).}

\item{labelValues}{Logical, indicating whether to label values (default is FALSE).}
}
\value{
A data frame containing the requested data.
}
\description{
This function retrieves data from REDCap using the REDCap API and an API token.
}
\details{
This function sends an HTTP POST request to the REDCap API to retrieve data.
Depending on the parameters, it can label both variables and values based on metadata.
}
\examples{
# Example usage:
data <- get_redcap_data(api_token = "your_api_token")

}
