% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_from_dictionary.R
\name{labels_from_dictionary}
\alias{labels_from_dictionary}
\title{Add variable labels to data from a data dictionary}
\usage{
labels_from_dictionary(
  data_df,
  dictionary_df,
  var_names_str,
  var_labels_str,
  var_choices_str = NULL,
  split_by = " \\\\| ",
  stata_path = NULL
)
}
\arguments{
\item{data_df}{The data frame to add labels to}

\item{dictionary_df}{The data dictionary containing variable names and data labels}

\item{var_names_str}{A string containing the name of the column in dictionary_df which contains the variable names}

\item{var_labels_str}{A string containing the name of the column in dictionary_df which contains the corresponding data labels}

\item{var_choices_str}{(Optional) A string containing the name of the column in dictionary_df which contains the response options}

\item{split_by}{(Optional, default = ' \\| ') The string to separate response options}

\item{stata_path}{(Optional) A string containing the file path of the location where the stata file should be saved}
}
\value{
The original data frame with labels assigned from the data dictionary provided
}
\description{
This function checks the data dictionary (dictionary_df) for any variable names which match those in the data (data_df). It then assigns the label from dictionary_df to the variables in data_df
}
\examples{
from_rcc <- labels_from_dictionary(data_df = export, dictionary_df = data_dictionary, var_names_str= "Variable...Field.Name", var_labels_str = "Field.Label")
}
