% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_missing_data.R
\name{missing_data_analysis}
\alias{missing_data_analysis}
\title{Calculate site-level missing data rates for required fields}
\usage{
missing_data_analysis(data, metadata)
}
\arguments{
\item{data}{A REDCap export dataset.}

\item{metadata}{A critical data item metadata dataframe.}
}
\value{
A dataframe containing missing counts and proportions by site and variable.
}
\description{
This function calculates missing data rates by site for required fields,
using expected event definitions from the metadata and excluding repeating instances.
}
\examples{
missing_results <- missing_data_analysis(data, metadata)

}
