% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_range_validation.R
\name{numeric_range_validation}
\alias{numeric_range_validation}
\title{Author: Paigan Aspinall
Date & version: 16FEB2024 V1.1.0
R version: 4.2.2}
\usage{
numeric_range_validation(dataset, rules)
}
\arguments{
\item{dataset}{A complete dataset.}

\item{rules}{A CSV file that is imported and contains all of the numeric range checks to be performed. Created using the template: NumericValidationTemplate_V1.0}
}
\value{
A dataset summarising potentially invalid numeric data.
}
\description{
Generate a dataset that returns all instances where the numeric validation rules defined in the CSV are broken.
}
\details{
Produces a dataset of potentially invalid numeric values.
}
\examples{
numeric_range_validation_output <- numeric_range_validation(dataset, rules)

}
