% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crf_completeness.R
\name{plot_crf_completeness}
\alias{plot_crf_completeness}
\title{Generate a stacked barchart to show timepoint completeness.}
\usage{
plot_crf_completeness(
  dataset,
  timepoint_names,
  category,
  api_token,
  test = FALSE
)
}
\arguments{
\item{timepoint_names}{A vector containing the list of timepoint names.}

\item{category}{The category by which data will be organised (e.g. "Site" or "Allocation").}

\item{api_token}{The API token for accessing REDCap.}

\item{test}{Logical, indicating whether to use the test environment (default is FALSE).}

\item{data}{A complete dataset.}
}
\value{
A dataset summarising event completeness.
}
\description{
This function produces a barchat that visualizes event completeness by allocation or site.
}
\examples{
timepoint_names <- c("week_0_blinded_ass_arm_1", "week_2_blinded_ass_arm_1", "week_6_blinded_ass_arm_1", "week_8_blinded_as_arm_1", "week_22_blinded_as_arm_1")

example_plot <- plot_crf_completeness(dataset, timepoint_names, "Site", your_api_token, test=TRUE)

}
