% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment_by_site_2.R
\name{plot_recruitment_by_site_2}
\alias{plot_recruitment_by_site_2}
\title{Generate a plot of recruitment at each site in a study. Updated version of plot_recruitment_by_site which allows sites to have different start dates.}
\usage{
plot_recruitment_by_site_2(
  recruitment_start,
  site_opening_dates,
  enrolment_date_data,
  site_data,
  site_names = NULL,
  recruitment_end = Sys.Date()
)
}
\arguments{
\item{recruitment_start}{A date which is the start date of recruitment, given in "2023-01-01" format.}

\item{site_opening_dates}{A vector where the opening date of each site is defined (e.g. site_opening_dates <- c("Plymouth"="2024-09-03", "Manchester"="2025-06-01"))}

\item{enrolment_date_data}{The column in the dataset where the enrolment date for each participant is defined, given in
format "dataset$enrolment_date".}

\item{site_data}{The column in the dataset where the participant's site is defined, given in format "dataset$site".}

\item{site_names}{A vector where the study sites can be defined as "c("Site 1", "Site2"), if left as NULL, these will be auto
populated from the site_data column.}

\item{recruitment_end}{A date which is the end date of recruitment, given in "2023-01-01" format. If left blank is auto
populated as today.}
}
\value{
A ggplot of recruitment at each site.
}
\description{
This function plots the recruitment at each site against the month and year; the sites can be defined or pulled from the data
the dates can be up to today or up until the end of recruitment.
}
