% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment_overall.R
\name{plot_recruitment_overall}
\alias{plot_recruitment_overall}
\title{Generate an actual versus target recruitment plot.}
\usage{
plot_recruitment_overall(
  data,
  recruitment_start,
  recruitment_length_months,
  recruitment_target,
  enrollment_date_column
)
}
\arguments{
\item{data}{A data frame to collect enrollment date data from.}

\item{recruitment_start}{A date which is the start date of recruitment, given in "2023-01-01" format.}

\item{recruitment_length_months}{A number which is the number of months that recruitment is open.}

\item{recruitment_target}{A number which is the target number of participants to be recruited.}

\item{enrollment_date_column}{A column which is where the enrollment date of participants is defined, given in "column_name" format.}
}
\value{
A ggplot of expected and actual recruitment.
}
\description{
This function plots the actual recruitment of a study and the target recruitment against the length of the study.
}
