% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_redcap_audit_actions.R
\name{plot_redcap_audit_actions}
\alias{plot_redcap_audit_actions}
\title{Create a line plot of record-related actions over time.}
\usage{
plot_redcap_audit_actions(
  audit_data,
  date_breaks,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{audit_data}{The REDCap dataset containing the audit log.}

\item{date_breaks}{Specifies the increments used on the x-axis. Can be set as "day" or "month".}

\item{start_date}{The date from which the graph should commence, optional.}

\item{end_date}{The date at which the graph should end, optional.}
}
\value{
A plot of actions by day.
}
\description{
This function uses audit log data to plot the number of actions taken over time.
}
\examples{
:
Example usage:

audit_plot <- plot_redcap_audit_actions(audit_data, date_breaks = "month", start_date= "2023-10-01", end_date=NULL )

}
