% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_lag.R
\name{plot_time_lag}
\alias{plot_time_lag}
\title{Produce a histograph of time lag.}
\usage{
plot_time_lag(
  dataset,
  lag_type,
  start_column,
  start_event,
  end_column,
  end_event,
  plot_name
)
}
\arguments{
\item{dataset}{Dataset from which time lag is to calculated.}

\item{lag_type}{Specifies whether the time lag will be calculated in days or months.}

\item{start_column}{Column where first date is specified.}

\item{start_event}{Name of event where first date is located.}

\item{end_column}{Column where second date is specified.}

\item{end_event}{Name of event where second event is specified.}

\item{plot_name}{Title of the plot.}
}
\value{
A histograph of difference between two dates.
}
\description{
This function calculate the difference between two specified date columns and plots a histogram of the differences.
}
\examples{
:
Example usage:

plot <- plot_time_lag(dataset, "months", "hoappointdat", "week_0_blinded_ass_arm_1", "hoappnextdat", "week_0_blinded_ass_arm_1", "Time until next scheduled appointment")

}
