% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_by_site.R
\name{recruitment_by_site}
\alias{recruitment_by_site}
\title{Generate a plot of recruitment at each site in a study.}
\usage{
recruitment_by_site(
  recruitment_start,
  enrollment_date_data,
  site_data,
  site_names = NULL,
  recruitment_end = Sys.Date()
)
}
\arguments{
\item{recruitment_start}{A date which is the start date of recruitment, given in "2023-01-01" format.}

\item{enrollment_date_data}{The column in the dataset where the enrollment date for each participant is defined, given in
format "dataset$enrollment_date".}

\item{site_data}{The column in the dataset where the participant's site is defined, given in format "dataset$site".}

\item{site_names}{A vector where the study sites can be defined as "c("Site 1", "Site2"), if left as NULL, these will be auto
populated from the site_data column.}

\item{recruitment_end}{A date which is the end date of recruitment, given in "2023-01-01" format. If left blank is auto
populated as today.}
}
\value{
A ggplot of recruitment at each site.
}
\description{
This function plots the recruitment at each site against the month and year; the sites can be defined or pulled from the data
the dates can be up to today or up until the end of recruitment.
}
