% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_fields.R
\name{remove_fields}
\alias{remove_fields}
\title{Removed multiple columns at a time from a dataset (for anonymisation).}
\usage{
remove_fields(dataset, fields_to_remove)
}
\arguments{
\item{dataset}{Dataset to beanonymised}

\item{fields_to_remove}{Vector defining the columns to be removed}
}
\value{
A dataset with defined columns removed
}
\description{
Removed multiple columns at a time from a dataset (for anonymisation).
}
\examples{
:
Example usage:

fields_to_remove <- c("ID", "DateOfBirth")
data_clean <- remove_fields(dataset, fields_to_remove)
}
