% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_withdrawn_events.R
\name{remove_withdrawn_events}
\alias{remove_withdrawn_events}
\title{Identify events that occur after withdrawal date and remove from dataset.}
\usage{
remove_withdrawn_events(
  dataset,
  withdrawal_event,
  withdrawal_date_field,
  scheduled_event_data,
  early_discontinuation = FALSE,
  early_discontinuation_field = NULL,
  early_discontinuation_value = NULL
)
}
\arguments{
\item{dataset}{A REDcap export dataset.}

\item{withdrawal_event}{The name of the event where withdrawals are reported.}

\item{withdrawal_date_field}{Field in the data where the withdrawal date is defined.}

\item{scheduled_event_data}{Dataset containing the paramaters for each event - template located in Github named "20240109_EventsTemplate.csv".}

\item{early_discontinuation}{Logical, can be set to TRUE if early discontinuations are also recorded in the withdrawal form. Default is FALSE.}

\item{early_discontinuation_field}{Field in the data where the user defines whether the withdrawal is an early discontinuation or not. Default is NULL.}

\item{early_discontinuation_value}{The value in the early discontinuation field that indicates the form is for a full withdrawal. Default is NULL.}
}
\value{
A dataframe without events scheduled to occur after withdrawal.
}
\description{
This function is used after add_missing_events and confirms that any added events are not expected to occur after withdrawal dates, any events occurring after withdrawal are removed.
}
\examples{
:
Example useage:

scheduled_event_data <- read.csv("20240110_ScheduledEvents_V1.0.csv")

complete_dataset <- remove_withdrawn_events(dataset, "withdraw_arm_1", "wiwithdrawdat", scheduled_event_data, early_discontinuation=TRUE, "wiearlydiscon", "Entire study")

}
