% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_with_date.R
\name{replace_with_date}
\alias{replace_with_date}
\title{Replace anchor_date with value from corresponding column}
\usage{
replace_with_date(row)
}
\arguments{
\item{row}{A named list or data frame row containing the data.}
}
\value{
The modified row with the 'anchor_date' column replaced if specified.
}
\description{
This function takes a row of data and replaces the value of the 'anchor_date'
column with the value from another column specified by the value in the
'anchor_date' column itself.
}
