% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_crf_completeness.R
\name{table_crf_completeness}
\alias{table_crf_completeness}
\title{Generate a table of CRF completeness at a specified time point.}
\usage{
table_crf_completeness(
  dataset,
  timepoint_name,
  category,
  api_token,
  test = FALSE
)
}
\arguments{
\item{timepoint_name}{The name of the timepoint for which we are measuring CRF completeness, e.g. "facetoface_screeni_arm_1".}

\item{category}{The category by which data will be organised, must be "Site" or "Allocation".}

\item{api_token}{The API token for accessing REDCap.}

\item{test}{Logical, indicating whether to use the test environment (default is FALSE).}

\item{data}{A complete dataset.}
}
\value{
A flextable summarising CRF completeness.
}
\description{
This function produces a flextable that summarises the number of complete and incomplete CRFs at a defined timepoint.
}
\examples{
timepoint_completeness_table <- table_crf_completeness(dataset, "facetoface_screeni_arm_1", "Site", your_api_token, test=FALSE)

}
