% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_crf_completeness_data_preparation.R
\name{table_crf_completeness_data_prep}
\alias{table_crf_completeness_data_prep}
\title{Generate a dataset from RCA data to be used in combination with the table_crf_completeness function.}
\usage{
table_crf_completeness_data_prep(
  dataset,
  timepoint_name,
  category,
  api_token,
  test = FALSE
)
}
\arguments{
\item{timepoint_name}{The column name of the timepoint for which we are measuring CRF completeness, e.g. "facetoface_screeni_arm_1".}

\item{category}{The category by which data will be organised, must be "Site" or "Allocation".}

\item{api_token}{The API token for accessing REDCap.}

\item{test}{Logical, indicating whether to use the test environment (default is FALSE).}

\item{data}{A complete dataset.}
}
\value{
A dataset ready for downstream processing.
}
\description{
This function produces datset ready to be processed into a crf completeness by timepoint summary figure.
}
\examples{
example_dataset<- table_crf_completeness_data_prep(dataset, "facetoface_screeni_arm_1", "Allocation", your_api_token, test=TRUE)

}
