% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_demographics_data_preparation.R
\name{table_demographics_data_preparation}
\alias{table_demographics_data_preparation}
\title{Prepare a dataset from which to plot a demographics summary table.}
\usage{
table_demographics_data_preparation(
  dataset,
  demographic_columns,
  demographics_event,
  dob_column = NULL,
  anchor_column = NULL,
  anchor_event = NULL
)
}
\arguments{
\item{dataset}{The dataset where the demographics data is being pulled from.}

\item{demographic_columns}{A vector of desired columns to be presented in the demographics table, separated by a comma.}

\item{demographics_event}{The name of the event in which demographics data is collected and where the demographic_columns are located.}

\item{dob_column}{The name of the date of birth column. NULL by default.}

\item{anchor_column}{The column location of the date from which age will be calculated (e.g. age at registration). NULL by default.}

\item{anchor_event}{The name of the event where the anchor_column is collected. NULL by default.}
}
\value{
A data frame containing the desired demographics data.
}
\description{
This function calculates the age at registration (if required) and isolates the desired demographics data into a new dataset.
}
\details{
This function calculated the age at the anchor_column date, if required, and then
selects all fo the desired demographics columns into a concise dataset.
}
