% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_query_summary_preparation.R
\name{table_query_summary_preparation}
\alias{table_query_summary_preparation}
\title{Produce a query dataset suitable for downstream use in flex tables and in con junction with plot_query_data}
\usage{
table_query_summary_preparation(query_data, DAGs = TRUE, start_date = NULL)
}
\arguments{
\item{query_data}{REDCap export of data resolution workflow.}

\item{DAGs}{Logical, to specify whether DAGs are used in the study (default is TRUE).}

\item{start_date}{Used to specify the date from which queries should be included in the output (default is NULL).}
}
\value{
A data frame containing the cleaned query data.
}
\description{
This function cleans the audit data from REDCap for downstream use.
}
\examples{
:
Example usage:

query_data <- read.csv("DataResolutionDashboard.csv")

queries <- table_query_summary_preparation(query_data, DAGs=TRUE, start_date = "2023-09-01")

}
