% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_sae_data_preparation.R
\name{table_sae_data_preparation}
\alias{table_sae_data_preparation}
\title{Produce a dataset of SAEs ready for use in a flextable}
\usage{
table_sae_data_preparation(
  dataset,
  sae_event_name,
  sae_date_column,
  sae_description_column,
  classification_columns,
  classification_text,
  report_outcome_column,
  resolution_outcome_column,
  outcome_date_columns,
  severity_column,
  expectedness_column,
  relatedness_column,
  allocation = TRUE,
  previous_report_date = NULL
)
}
\arguments{
\item{dataset}{A REDCap dataset.}

\item{sae_event_name}{Name of the SAE event.}

\item{sae_date_column}{Column where the date of the SAE is recorded.}

\item{sae_description_column}{Column where the SAE is described.}

\item{classification_columns}{A vector of the SAE classification columns.}

\item{classification_text}{A vector of the classification descriptions, in the order they appear in the classification_columns vector.}

\item{report_outcome_column}{Column where the outcome at the time of reporting is recorded.}

\item{resolution_outcome_column}{Column where the outcome at the time of resolution is recorded.}

\item{outcome_date_columns}{A vector of column names where the outcome date(s) are recorded.}

\item{severity_column}{Column where the severity of the SAE is recorded.}

\item{expectedness_column}{Column where the expectedness of the SAE is recorded.}

\item{relatedness_column}{Column where the relatedness of the SAE is recorded.}

\item{allocation}{Logical, determines whether allocation is included in the table (default is TRUE).}

\item{previous_report_date}{Date from which the SAEs will be reported in the table (default is NULL).}
}
\value{
A dataset containing SAE data.
}
\description{
This function cleans the SAE data and produces a dataframe.
}
\examples{
# Example usage:

classification_columns <- c("aeclass___1", "aeclass___2", "aeclass___3", "aeclass___4", "aeclass___5")
classification_text <- c("Resulted in death", "Life-threatening event", "Prolonged/required hospitalisation", "Persistent/significant disability/incapacity", "Significant or important medical event")
outcome_date_columns <- c("aerecover1dat", "aedeath1dat", "aerecover2dat", "aedeath2dat")

sae_table <- table_sae_data_preparation(your_dataset, "sae_reporting_arm_1", "aeonsetdat", "aedescribeft", classification_columns, 
                                         classification_text, "aeoutcome1", "aeoutcome2", outcome_date_columns, "aepiseverity",
                                         "aeciexpect", "aepitrialyn", allocation = TRUE, "2023-09-01")

}
