% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_withdrawals_data_preparation.R
\name{table_withdrawals_data_preparation}
\alias{table_withdrawals_data_preparation}
\title{Prepare a dataset from which to plot a withdrawals table.}
\usage{
table_withdrawals_data_preparation(
  dataset,
  withdrawal_date_column,
  withdrawal_reason_column,
  withdrawal_event_name,
  allocation = FALSE,
  early_discontinuation_column = NULL,
  full_withdrawal_option = NULL
)
}
\arguments{
\item{dataset}{The dataset where the withdrawal data is being pulled from.}

\item{withdrawal_date_column}{The column in the dataset where the date of the withdrawal is specified.}

\item{withdrawal_reason_column}{The column in the dataset where the reason for the withdrawal is specified.}

\item{withdrawal_event_name}{The name of the event where withdrawals are recorded.}

\item{allocation}{Logical, indicating whether the allocation data should be included in the output. Default is FALSE.}

\item{early_discontinuation_column}{Name of the column where whether the withdrawal is an early discontinuation is specified. Default is NULL.}

\item{full_withdrawal_option}{The response option for early_discontinuation_column that is selected for a full withdrawal only. Default is NULL.}
}
\value{
A data frame containing withdrawal data.
}
\description{
This function removed early discontinuations (if required) and selects only the required withdrawal data.
}
\details{
This function removes early discontinuation data (if required) and selects only the required data
from the dataset to produce a withdrawal table.
}
\examples{
# Example usage:
withdrawal_table_data <- table_withdrawals_data_preparation(dataset, "wiwithdrawdat", "wireason", "withdraw_arm_1", allocation = TRUE,
                          "wiearlydiscon", "Entire study")
Example downstream usage: 
withdrawals_table <- flextable(withdrawal_table_data) \%>\%
 set_table_properties(width = .75, layout = "autofit")
 
 @export
}
