% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_data_cleaning.R
\name{allocation_data_cleaning}
\alias{allocation_data_cleaning}
\title{Apply coded allocations to every row}
\usage{
allocation_data_cleaning(
  data,
  data_dictionary,
  field_name_column,
  select_choices_column,
  allocation_column,
  allocation_event_name
)
}
\arguments{
\item{data}{A REDcap export dataset.}

\item{data_dictionary}{A REDcap export data dictionary.}

\item{field_name_column}{Column name where field names are stored in data dictionary}

\item{select_choices_column}{Column name where allocation data codes (1 for Control, 2 for Intervention) will be retrieved}

\item{allocation_column}{Column name where allocation data is stored.}

\item{allocation_event_name}{Event name where allocations are assigned to participants.}
}
\value{
A dataframe containing allocations.
}
\description{
This function applies allocations to all rows for a participant
}
\examples{
:
Example useage:
dataset <- allocation_data_cleaning(data, data_dictionary, "Variable...Field.Name", "Choices..Calculations..OR.Slider.Labels", "pdallocation", "randomisation_arm_1")


}
