% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_JSON_from_column.R
\name{extract_JSON_from_column}
\alias{extract_JSON_from_column}
\title{Extract JSON from dataframe column}
\usage{
extract_JSON_from_column(data, col_name)
}
\arguments{
\item{data}{A dataset with a column that contains only JSON objects}
}
\value{
The input dataset with the JSON column unnested and each JSON property
}
\description{
Takes a dataframe with a JSON object in one of its columns and unnests the JSON so that each JSON property becomes its own unique column. Also extracts values corresponding to each row.
}
\examples{
example1 <- extract_JSON_from_column(data, "column_name_as_string");
}
