% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_duplicate_detection.R
\name{similarity_detection}
\alias{similarity_detection}
\title{Detect duplicate or highly similar records}
\usage{
similarity_detection(
  data,
  metadata,
  min_comparable_fields = 3,
  similarity_threshold = 0.8
)
}
\arguments{
\item{data}{A REDCap export dataset.}

\item{metadata}{A critical data item metadata dataframe.}

\item{min_comparable_fields}{Minimum number of shared non-missing fields required for comparison.}

\item{similarity_threshold}{Similarity threshold used to flag record pairs.}
}
\value{
A dataframe containing pairs of records with high similarity.
}
\description{
This function compares non-repeating REDCap records across required critical
data items and identifies pairs of records with high similarity.
}
\examples{
similarity_results <- similarity_detection(data, metadata)

}
