% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_demographics_data_preparation.R
\name{table_demographics_data_preparation}
\alias{table_demographics_data_preparation}
\title{Prepare a dataset from which to plot a demographics summary table.}
\usage{
table_demographics_data_preparation(
  dataset,
  demographic_columns,
  demographics_event,
  category_column,
  dob_column = NULL,
  anchor_column = NULL,
  anchor_event = NULL
)
}
\arguments{
\item{dataset}{The dataset where the demographics data is being pulled from.}

\item{demographic_columns}{A vector of desired columns to be presented in the demographics table, separated by a comma.}

\item{demographics_event}{The name of the event in which demographics data is collected and where the demographic_columns are located.}

\item{category_column}{The name of the column that we organise the final dataset by}

\item{dob_column}{The name of the date of birth column. NULL by default.}

\item{anchor_column}{The column location of the date from which age will be calculated (e.g. age at registration). NULL by default.}

\item{anchor_event}{The name of the event where the anchor_column is collected. NULL by default.}
}
\value{
A data frame containing the desired demographics data.
}
\description{
This function calculates the age at registration (if required) and isolates the desired demographics data into a new dataset.
}
\details{
This function calculated the age at the anchor_column date, if required, and then
selects all fo the desired demographics columns into a concise dataset.
}
\examples{
# Example usage:

demographic_columns <- c("dmsexatbirth", "dmethnic", "dmemployment", "dmlivewith", "dmmarital", "dmindependence")

demographics_data <- table_demographics_data_preparation(demo_dataset, demographics_columns,'Face-to-face screening', ‘Allocation’, dob_column = 'dmyob',
                                                         anchor_column = 'pdregisterdat', anchor_event = 'Participant identification')
#example downstream usage:
baseline_characteristics <-  demographics_table_data \%>\% 
tbl_summary(by = category,
            label = list(Age ~ "Age", dmsexatbirth ~ "Sex at birth", dmethnic ~ "Ethnicity",
                         dmemployment ~ "Employment status", dmlivewith ~ "Number of cohabitants",
                         dmmarital ~ "Marital status", dmindependence ~ "Independence level"),
            type = list(Age ~ "continuous", dmsexatbirth ~ "categorical", dmethnic ~ "categorical",
                        dmemployment ~ "categorical", dmlivewith ~ "categorical",
                        dmmarital ~ "categorical", dmindependence ~ "categorical"),
            statistic = list(all_categorical() ~ "{n} / {N} ({p}\%)",
                             all_continuous() ~ "{mean} ({sd})")) \%>\%
  italicize_levels() \%>\%
  bold_labels() \%>\%
  modify_caption("Participant Characteristics") \%>\%
  as_flex_table()


}
